Cleaning FormulaGrader and IntegralGrader
