+++ tests/formulagrader/test_integralgrader.py
@@ -515,8 +515,57 @@ def test_debug_message():
         "sin(x)<br/>\n"
         "x<br/>\n"
         "<br/>\n"
+        "==============================================================<br/>\n"
+        "IntegralGrader Debug Info<br/>\n"
+        "==============================================================<br/>\n"
+        "Functions available during evaluation and allowed in answer:<br/>\n"
+        "{{{u}'abs': <function absolute at 0x...>,<br/>\n"
+        " {u}'arccos': <function arccos at 0x...>,<br/>\n"
+        " {u}'arccosh': <function arccosh at 0x...>,<br/>\n"
+        " {u}'arccot': <function arccot at 0x...>,<br/>\n"
+        " {u}'arccoth': <function arccoth at 0x...>,<br/>\n"
+        " {u}'arccsc': <function arccsc at 0x...>,<br/>\n"
+        " {u}'arccsch': <function arccsch at 0x...>,<br/>\n"
+        " {u}'arcsec': <function arcsec at 0x...>,<br/>\n"
+        " {u}'arcsech': <function arcsech at 0x...>,<br/>\n"
+        " {u}'arcsin': <function arcsin at 0x...>,<br/>\n"
+        " {u}'arcsinh': <function arcsinh at 0x...>,<br/>\n"
+        " {u}'arctan': <function arctan at 0x...>,<br/>\n"
+        " {u}'arctan2': <function arctan2 at 0x...>,<br/>\n"
+        " {u}'arctanh': <function arctanh at 0x...>,<br/>\n"
+        " {u}'ceil': <function ceil at 0x...>,<br/>\n"
+        " {u}'conj': <ufunc 'conjugate'>,<br/>\n"
+        " {u}'cos': <function cos at 0x...>,<br/>\n"
+        " {u}'cosh': <function cosh at 0x...>,<br/>\n"
+        " {u}'cot': <function cot at 0x...>,<br/>\n"
+        " {u}'coth': <function coth at 0x...>,<br/>\n"
+        " {u}'csc': <function csc at 0x...>,<br/>\n"
+        " {u}'csch': <function csch at 0x...>,<br/>\n"
+        " {u}'exp': <function exp at 0x...>,<br/>\n"
+        " {u}'fact': <function factorial at 0x...>,<br/>\n"
+        " {u}'factorial': <function factorial at 0x...>,<br/>\n"
+        " {u}'floor': <function floor at 0x...>,<br/>\n"
+        " {u}'im': <function imag at 0x...>,<br/>\n"
+        " {u}'kronecker': <function kronecker at 0x...>,<br/>\n"
+        " {u}'ln': <function log at 0x...>,<br/>\n"
+        " {u}'log10': <function log10 at 0x...>,<br/>\n"
+        " {u}'log2': <function log2 at 0x...>,<br/>\n"
+        " {u}'max': <function max at 0x...>,<br/>\n"
+        " {u}'min': <function min at 0x...>,<br/>\n"
+        " {u}'re': <function real at 0x...>,<br/>\n"
+        " {u}'sec': <function sec at 0x...>,<br/>\n"
+        " {u}'sech': <function sech at 0x...>,<br/>\n"
+        " {u}'sin': <function sin at 0x...>,<br/>\n"
+        " {u}'sinh': <function sinh at 0x...>,<br/>\n"
+        " {u}'sqrt': <function sqrt at 0x...>,<br/>\n"
+        " {u}'tan': <function tan at 0x...>,<br/>\n"
+        " {u}'tanh': <function tanh at 0x...>}}<br/>\n"
+        "Functions available during evaluation and disallowed in answer:<br/>\n"
+        "{{}}<br/>\n"
+        "<br/>\n"
+        "<br/>\n"
         "==============================================<br/>\n"
-        "Integration Data for Sample Number 0<br/>\n"
+        "Integration Data for Sample Number 1 of 2<br/>\n"
         "==============================================<br/>\n"
         "Variables: {{{u}'e': 2.718281828459045,<br/>\n"
         " {u}'i': 1j,<br/>\n"
@@ -546,7 +595,7 @@ def test_debug_message():
         "<br/>\n"
         "<br/>\n"
         "==============================================<br/>\n"
-        "Integration Data for Sample Number 1<br/>\n"
+        "Integration Data for Sample Number 2 of 2<br/>\n"
         "==============================================<br/>\n"
         "Variables: {{{u}'e': 2.718281828459045,<br/>\n"
         " {u}'i': 1j,<br/>\n"
@@ -573,6 +622,15 @@ def test_debug_message():
         "Numerical Value: None<br/>\n"
         "Error Estimate: None<br/>\n"
         "Number of integrand evaluations: None<br/>\n"
+        "<br/>\n"
+        "<br/>\n"
+        "==========================================<br/>\n"
+        "Comparison Data for All 2 Samples<br/>\n"
+        "==========================================<br/>\n"
+        "Comparer Function: EqualityComparer({{{u}'transform': <function identity_transform at 0x...>}})<br/>\n"
+        "Comparison Results:<br/>\n"
+        "[{{{u}'grade_decimal': 1.0, {u}'msg': {u}'', {u}'ok': True}},<br/>\n"
+        " {{{u}'grade_decimal': 1.0, {u}'msg': {u}'', {u}'ok': True}}]<br/>\n"
         "</pre>"
     ).format(version=__version__, u=UNICODE_PREFIX)
     expected_message = round_decimals_in_string(expected_message)
@@ -583,15 +641,14 @@ def test_debug_message():
 
 def test_debug_message_complex_integrand():
     grader = IntegralGrader(
-    complex_integrand=True,
+        complex_integrand=True,
         answers={
             'lower': '-2',
             'upper': '4',
             'integrand': 'cos(s) + i*sin(s)',
             'integration_variable': 's'
         },
-        debug=True,
-        samples=2
+        debug=True
     )
     student_input = ['-1', '5', 'sqrt(x)', 'x']
     expected_message = (
@@ -602,8 +659,57 @@ def test_debug_message_complex_integrand():
         "sqrt(x)<br/>\n"
         "x<br/>\n"
         "<br/>\n"
+        "==============================================================<br/>\n"
+        "IntegralGrader Debug Info<br/>\n"
+        "==============================================================<br/>\n"
+        "Functions available during evaluation and allowed in answer:<br/>\n"
+        "{{{u}'abs': <function absolute at 0x...>,<br/>\n"
+        " {u}'arccos': <function arccos at 0x...>,<br/>\n"
+        " {u}'arccosh': <function arccosh at 0x...>,<br/>\n"
+        " {u}'arccot': <function arccot at 0x...>,<br/>\n"
+        " {u}'arccoth': <function arccoth at 0x...>,<br/>\n"
+        " {u}'arccsc': <function arccsc at 0x...>,<br/>\n"
+        " {u}'arccsch': <function arccsch at 0x...>,<br/>\n"
+        " {u}'arcsec': <function arcsec at 0x...>,<br/>\n"
+        " {u}'arcsech': <function arcsech at 0x...>,<br/>\n"
+        " {u}'arcsin': <function arcsin at 0x...>,<br/>\n"
+        " {u}'arcsinh': <function arcsinh at 0x...>,<br/>\n"
+        " {u}'arctan': <function arctan at 0x...>,<br/>\n"
+        " {u}'arctan2': <function arctan2 at 0x...>,<br/>\n"
+        " {u}'arctanh': <function arctanh at 0x...>,<br/>\n"
+        " {u}'ceil': <function ceil at 0x...>,<br/>\n"
+        " {u}'conj': <ufunc 'conjugate'>,<br/>\n"
+        " {u}'cos': <function cos at 0x...>,<br/>\n"
+        " {u}'cosh': <function cosh at 0x...>,<br/>\n"
+        " {u}'cot': <function cot at 0x...>,<br/>\n"
+        " {u}'coth': <function coth at 0x...>,<br/>\n"
+        " {u}'csc': <function csc at 0x...>,<br/>\n"
+        " {u}'csch': <function csch at 0x...>,<br/>\n"
+        " {u}'exp': <function exp at 0x...>,<br/>\n"
+        " {u}'fact': <function factorial at 0x...>,<br/>\n"
+        " {u}'factorial': <function factorial at 0x...>,<br/>\n"
+        " {u}'floor': <function floor at 0x...>,<br/>\n"
+        " {u}'im': <function imag at 0x...>,<br/>\n"
+        " {u}'kronecker': <function kronecker at 0x...>,<br/>\n"
+        " {u}'ln': <function log at 0x...>,<br/>\n"
+        " {u}'log10': <function log10 at 0x...>,<br/>\n"
+        " {u}'log2': <function log2 at 0x...>,<br/>\n"
+        " {u}'max': <function max at 0x...>,<br/>\n"
+        " {u}'min': <function min at 0x...>,<br/>\n"
+        " {u}'re': <function real at 0x...>,<br/>\n"
+        " {u}'sec': <function sec at 0x...>,<br/>\n"
+        " {u}'sech': <function sech at 0x...>,<br/>\n"
+        " {u}'sin': <function sin at 0x...>,<br/>\n"
+        " {u}'sinh': <function sinh at 0x...>,<br/>\n"
+        " {u}'sqrt': <function sqrt at 0x...>,<br/>\n"
+        " {u}'tan': <function tan at 0x...>,<br/>\n"
+        " {u}'tanh': <function tanh at 0x...>}}<br/>\n"
+        "Functions available during evaluation and disallowed in answer:<br/>\n"
+        "{{}}<br/>\n"
+        "<br/>\n"
+        "<br/>\n"
         "==============================================<br/>\n"
-        "Integration Data for Sample Number 0<br/>\n"
+        "Integration Data for Sample Number 1 of 1<br/>\n"
         "==============================================<br/>\n"
         "Variables: {{{u}'e': 2.718281828459045,<br/>\n"
         " {u}'i': 1j,<br/>\n"
@@ -630,6 +736,14 @@ def test_debug_message_complex_integrand():
         "Numerical Value: 0.237496784316<br/>\n"
         "Error Estimate: 4.1882074502e-14<br/>\n"
         "Number of integrand evaluations: 21<br/>\n"
+        "<br/>\n"
+        "<br/>\n"
+        "==========================================<br/>\n"
+        "Comparison Data for All 1 Samples<br/>\n"
+        "==========================================<br/>\n"
+        "Comparer Function: EqualityComparer({{{u}'transform': <function identity_transform at 0x...>}})<br/>\n"
+        "Comparison Results:<br/>\n"
+        "[{{{u}'grade_decimal': 0, {u}'msg': {u}'', {u}'ok': False}}]<br/>\n"
         "</pre>"
     ).format(version=__version__, u=UNICODE_PREFIX)
     expected_message = round_decimals_in_string(expected_message)
